found bubbly word :)

+++ src/FindProductClass.java
@@ -5,7 +5,7 @@ import java.util.Scanner;
  */
 public class FindProductClass {
     public static void main(String[] args) {
-        int answer=1;
+        double answer=1;
         Scanner sc = new Scanner(System.in);
         int n=sc.nextInt();
         int [] array= new int[n];
@@ -13,7 +13,7 @@ public class FindProductClass {
             array[i]=sc.nextInt();
         }
         for (int i=0;i<array.length;i++){
-            answer= (int) (answer*array[i]%(Math.pow(10,9)+7));
+            answer=  (answer*array[i]%(Math.pow(10,9)+7));
         }
         System.out.println(answer);
     }
